Move monitor / app tests into platform router module
